@@ -12,93 +12,70 @@
   "homepage": "https://hasura.io/",
   "keywords": [],
   "scripts": {
-    "start": "concurrently --kill-others \"npm run start-prod\"",
-    "start-prod": "better-npm-run start-prod",
+    "start-dev": "node -r dotenv/config ./bin/server.js",
+    "watch-client": "node -r dotenv/config webpack/webpack-dev-server.js",
+    "dev": "concurrently --kill-others \"npm run watch-client\" \"npm run start-dev\"",
+    "start-prod": "cross-env NODE_PATH=./src PORT=8080 NODE_ENV=production node ./bin/server.js",
     "build": "webpack --progress -p --colors --display-error-details --config webpack/prod.config.js",
     "server-build": "make server-build",
     "build-unused": "webpack --verbose --colors --display-error-details --config webpack/prod.config.js --json | webpack-unused -s src",
     "postinstall": "webpack --display-error-details --config webpack/prod.config.js",
-    "lint": "eslint -c .eslintrc src --ext .js,.ts,.tsx",
-    "start-dev": "better-npm-run start-dev",
-    "watch-client": "better-npm-run watch-client",
-    "dev": "concurrently --kill-others \"npm run watch-client\" \"npm run start-dev\" ",
     "cypress": "cypress open",
     "test": "cypress run --spec 'cypress/integration/**/**/test.ts' --key $CYPRESS_KEY --parallel --record",
-    "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx}\""
+    "lint": "eslint -c .eslintrc src --ext .js,.ts,.tsx",
+    "lint:quiet": "eslint --quiet -c .eslintrc src --ext .js,.ts,.tsx",
+    "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx}\"",
+    "format:check": "prettier --check \"src/**/*.{js,jsx,ts,tsx}\""
   },
   "lint-staged": {
     "*.{js,ts,tsx}": [
-      "eslint --fix",
-      "git add"
+      "eslint --fix"
     ],
     "*.{js,json,ts,tsx,css,md}": [
-      "prettier --write",
-      "git add"
+      "prettier --write"
     ]
-  },
-  "betterScripts": {
-    "start-prod": {
-      "command": "node ./bin/server.js",
-      "env": {
-        "NODE_PATH": "./src",
-        "NODE_ENV": "production",
-        "PORT": 8080
-      }
-    },
-    "start-dev": {
-      "command": "node -r dotenv/config ./bin/server.js"
-    },
-    "watch-client": {
-      "command": "node -r dotenv/config webpack/webpack-dev-server.js"
-    }
   },
   "prettier": {
     "singleQuote": true,
     "trailingComma": "es5",
     "printWidth": 80,
-    "tabWidth": 2
+    "tabWidth": 2,
+    "arrowParens": "avoid"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged"
+      "pre-commit": "node ./scripts/preCommit.js"
     }
   },
   "dependencies": {
-    "@graphql-codegen/core": "1.13.0",
-    "@graphql-codegen/typescript": "1.13.0",
-    "apollo-link": "1.2.13",
-    "apollo-link-ws": "1.0.19",
+    "@graphql-codegen/core": "1.13.5",
+    "@graphql-codegen/typescript": "1.13.5",
+    "ace-builds": "^1.4.11",
+    "apollo-link": "1.2.14",
+    "apollo-link-ws": "1.0.20",
     "brace": "0.11.1",
-    "deep-equal": "1.1.1",
     "graphiql": "1.0.0-alpha.0",
     "graphiql-explorer": "0.6.2",
     "graphql": "14.5.8",
     "graphql-voyager": "1.0.0-rc.29",
     "history": "3.3.0",
-    "hoist-non-react-statics": "1.2.0",
     "inflection": "1.12.0",
-    "invariant": "2.2.4",
     "isomorphic-fetch": "2.2.1",
     "jsonwebtoken": "8.5.1",
     "less": "3.11.1",
-    "lru-memoize": "1.0.0",
-    "map-props": "1.0.0",
-    "match-sorter": "2.3.0",
-    "multireducer": "1.0.2",
+    "moment": "^2.26.0",
     "piping": "0.3.2",
-    "pretty-error": "1.2.0",
     "prop-types": "15.7.2",
-    "query-string": "6.11.1",
-    "react": "16.8.6",
+    "react": "16.13.1",
     "react-ace": "8.0.0",
-    "react-autosuggest": "9.4.3",
+    "react-autosuggest": "10.0.2",
     "react-bootstrap": "0.32.4",
     "react-copy-to-clipboard": "5.0.2",
-    "react-dom": "16.8.6",
+    "react-datetime": "^2.16.3",
+    "react-dom": "16.13.1",
     "react-helmet": "5.2.1",
     "react-icons": "3.9.0",
     "react-modal": "3.11.2",
-    "react-notification-system": "0.2.17",
     "react-notification-system-redux": "1.2.0",
     "react-progress-bar-plus": "1.3.1",
     "react-redux": "5.1.2",
