@@ -19,6 +19,7 @@ const Permissions = ({
   permissionEdit,
   isEditing,
   isFetching,
+  readOnlyMode = false,
 }) => {
   React.useEffect(() => {
     dispatch(fetchRoleList());
@@ -79,7 +80,6 @@ const Permissions = ({
             dispatch(permCloseEdit());
           };
 
-          const isEditAllowed = role !== 'admin';
           const isCurrEdit =
             isEditing &&
             (permissionEdit.role === role ||
@@ -87,14 +87,15 @@ const Permissions = ({
           let editIcon = '';
           let className = '';
           let onClick = () => {};
-          if (isEditAllowed) {
+          if (role !== 'admin' && !readOnlyMode) {
             editIcon = getEditIcon();
 
-            className += styles.clickableCell;
-            onClick = dispatchOpenEdit(queryType);
             if (isCurrEdit) {
               onClick = dispatchCloseEdit;
               className += ` ${styles.currEdit}`;
+            } else {
+              className += styles.clickableCell;
+              onClick = dispatchOpenEdit(queryType);
             }
           }
 
@@ -169,12 +170,14 @@ const Permissions = ({
       />
       {getPermissionsTable()}
       <div className={`${styles.add_mar_bottom}`}>
-        <PermissionEditor
-          permissionEdit={permissionEdit}
-          dispatch={dispatch}
-          isFetching={isFetching}
-          isEditing={isEditing}
-        />
+        {!readOnlyMode && (
+          <PermissionEditor
+            permissionEdit={permissionEdit}
+            dispatch={dispatch}
+            isFetching={isFetching}
+            isEditing={isEditing}
+          />
+        )}
       </div>
     </div>
   );
