@@ -189,18 +189,22 @@ const RelationshipEditor = ({
           onChange={setRelRefSchema}
           disabled={!name}
         >
-          {// default unselected option
-          refSchema === '' && (
-            <option value={''} disabled>
-              {'-- reference schema --'}
-            </option>
-          )}
-          {// all reference schema options
-          orderedSchemaList.map((rs, j) => (
-            <option key={j} value={rs}>
-              {rs}
-            </option>
-          ))}
+          {
+            // default unselected option
+            refSchema === '' && (
+              <option value={''} disabled>
+                {'-- reference schema --'}
+              </option>
+            )
+          }
+          {
+            // all reference schema options
+            orderedSchemaList.map((rs, j) => (
+              <option key={j} value={rs}>
+                {rs}
+              </option>
+            ))
+          }
         </select>
       </div>
     );
@@ -370,7 +374,7 @@ const RelationshipEditor = ({
 };
 
 const RelEditor = props => {
-  const { dispatch, relConfig, objectType, isNew } = props;
+  const { dispatch, relConfig, objectType, isNew, readOnlyMode } = props;
 
   const [relConfigState, setRelConfigState] = React.useState(null);
 
@@ -448,6 +452,7 @@ const RelEditor = props => {
       removeFunc={removeFunc}
       expandButtonText={expandButtonText}
       collapseButtonText={collapseButtonText}
+      readOnlyMode={readOnlyMode}
     />
   );
 };
