@@ -4,7 +4,11 @@ import { Link } from 'react-router';
 import LeftSubSidebar from '../../../Common/Layout/LeftSubSidebar/LeftSubSidebar';
 import styles from '../../../Common/Layout/LeftSubSidebar/LeftSubSidebar.scss';
 
-const LeftSidebar = ({ appPrefix, common: { actions, currentAction } }) => {
+const LeftSidebar = ({
+  appPrefix,
+  common: { actions, currentAction },
+  readOnlyMode,
+}) => {
   const [searchText, setSearchText] = React.useState('');
 
   const handleSearch = e => setSearchText(e.target.value);
@@ -81,7 +85,7 @@ const LeftSidebar = ({ appPrefix, common: { actions, currentAction } }) => {
 
   return (
     <LeftSubSidebar
-      showAddBtn
+      showAddBtn={!readOnlyMode}
       searchInput={getSearchInput()}
       heading={`Actions (${actionsList.length})`}
       addLink={`${appPrefix}/manage/add`}
