@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Connect } from 'react-redux';
-import { LocationShape } from 'react-router/lib/PropTypes';
+import { RouteComponentProps } from 'react-router';
 import Sidebar from './Sidebar';
 import PageContainer from '../../Common/Layout/PageContainer/PageContainer';
 
@@ -11,7 +11,7 @@ type Metadata = {
 };
 
 type ExternalProps = {
-  location: LocationShape;
+  location: RouteComponentProps<{}, {}>['location'];
   children: JSX.Element;
 };
 
